backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==22.0.0
idna==3.7
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pypng==0.20220715.0
pyshorteners==1.0.1
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.31.0
shortuuid==1.0.13
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
